+++ src/main/java/com/braintreegateway/PaymentMethodNonceDetails.java
@@ -14,6 +14,7 @@ public class PaymentMethodNonceDetails {
     private String paymentInstrumentName;
     private String username;
     private String venmoUserId;
+    private PaymentMethodNonceDetailsPayerInfo payerInfo;
 
     public PaymentMethodNonceDetails(NodeWrapper node) {
         cardType = node.findString("card-type");
@@ -26,19 +27,29 @@ public class PaymentMethodNonceDetails {
         paymentInstrumentName = node.findString("payment-instrument-name");
         username = node.findString("username");
         venmoUserId = node.findString("venmo-user-id");
+
+        NodeWrapper payerInfoNode = node.findFirst("payer-info");
+        if (payerInfoNode != null && !payerInfoNode.isBlank()) {
+            payerInfo = new PaymentMethodNonceDetailsPayerInfo(payerInfoNode);
+        }
     }
 
-    public PaymentMethodNonceDetails(Map<String, String> map) {
-        cardType = map.get("card-type");
-        cardholderName = map.get("cardholder-name");
-        dpanLastTwo = map.get("dpan-last-two");
-        email = map.get("email");
-        bin = map.get("bin");
-        lastFour = map.get("last-four");
-        lastTwo = map.get("last-two");
-        paymentInstrumentName = map.get("payment-instrument-name");
-        username = map.get("username");
-        venmoUserId = map.get("venmo-user-id");
+    public PaymentMethodNonceDetails(Map<String, Object> map) {
+        cardType = (String) map.get("card-type");
+        cardholderName = (String) map.get("cardholder-name");
+        dpanLastTwo = (String) map.get("dpan-last-two");
+        email = (String) map.get("email");
+        bin = (String) map.get("bin");
+        lastFour = (String) map.get("last-four");
+        lastTwo = (String) map.get("last-two");
+        paymentInstrumentName = (String) map.get("payment-instrument-name");
+        username = (String) map.get("username");
+        venmoUserId = (String) map.get("venmo-user-id");
+
+        Map<String, String> payerInfoMap = (Map) map.get("payer-info");
+        if (payerInfoMap != null) {
+            payerInfo = new PaymentMethodNonceDetailsPayerInfo(payerInfoMap);
+        }
     }
 
     public String getCardType() {
@@ -79,5 +90,9 @@ public class PaymentMethodNonceDetails {
 
     public String getVenmoUserId() {
         return venmoUserId;
+    }
+
+    public PaymentMethodNonceDetailsPayerInfo getPayerInfo() {
+        return payerInfo;
     }
 }